+++ test/entrypoint/srv_select.lua
@@ -51,10 +51,32 @@ package.preload['customers-storage'] = function()
             })
             customers_space:create_index('full_name', {
                 parts = {
-                    { field = 'name', collation = 'unicode_ci' },
-                    { field = 'last_name', collation = 'unicode_ci' },
+                    {field = 'name', collation = 'unicode_ci'},
+                    {field = 'last_name', collation = 'unicode_ci'} ,
                 },
                 unique = false,
+                if_not_exists = true,
+            })
+
+            local coord_space = box.schema.space.create('coord', {
+                format = {
+                      {name = 'x', type = 'unsigned'},
+                      {name = 'y', type = 'unsigned'},
+                      {name = 'bucket_id', type = 'unsigned'},
+                },
+                if_not_exists = true,
+                engine = engine,
+            })
+            -- primary index
+            coord_space:create_index('primary', {
+                parts = {
+                    {field = 'x'},
+                    {field = 'y'},
+                },
+                if_not_exists = true,
+            })
+            coord_space:create_index('bucket_id', {
+                parts = { {field = 'bucket_id'} },
                 if_not_exists = true,
             })
         end,